Add a utility to dynamically find all provided IteratorTestCases.
