+++ iterator-test-harness/src/test/java/org/apache/accumulo/iteratortest/WholeRowIteratorTest.java
@@ -18,7 +18,6 @@ package org.apache.accumulo.iteratortest;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map.Entry;
@@ -29,8 +28,7 @@ import org.apache.accumulo.core.data.Range;
 import org.apache.accumulo.core.data.Value;
 import org.apache.accumulo.core.iterators.user.WholeRowIterator;
 import org.apache.accumulo.iteratortest.junit4.BaseJUnit4IteratorTest;
-import org.apache.accumulo.iteratortest.testcases.InstantiationTestCase;
-import org.apache.accumulo.iteratortest.testcases.MultipleHasTopCalls;
+import org.apache.accumulo.iteratortest.testcases.IteratorTestCase;
 import org.apache.hadoop.io.Text;
 import org.junit.runners.Parameterized.Parameters;
 
@@ -43,7 +41,7 @@ public class WholeRowIteratorTest extends BaseJUnit4IteratorTest {
   public static Object[][] parameters() {
     IteratorTestInput input = getIteratorInput();
     IteratorTestOutput output = getIteratorOutput();
-    List<IteratorTestCase> tests = Arrays.<IteratorTestCase> asList(new MultipleHasTopCalls(), new InstantiationTestCase());
+    List<IteratorTestCase> tests = IteratorTestCaseFinder.findAllTestCases();
     return BaseJUnit4IteratorTest.createParameters(input, output, tests);
   }
