+++ pom.xml
@@ -527,6 +527,11 @@
         <version>${powermock.version}</version>
       </dependency>
       <dependency>
+      <groupId>org.reflections</groupId>
+        <artifactId>reflections</artifactId>
+        <version>0.9.10</version>
+      </dependency>
+      <dependency>
         <groupId>org.slf4j</groupId>
         <artifactId>slf4j-api</artifactId>
         <version>${slf4j.version}</version>
